backend/requirements.txt
# Flask
Flask>=3.0.0
Flask-SQLAlchemy
SQLAlchemy>=2.0.0
Flask-Migrate
alembic
python-dotenv
bcrypt
pytz
requests
sentry-sdk
python-json-logger
gunicorn
pytest
psycopg2-binary
